Bowling: wrap paragraphs and fix some minor grammar
The paragraphs rambled without newlines. They are now line wrapped to match
the other exercises.

There was a missing article 'a' in the beginning. The phrasing "after the
throw" is a little confusing. Does this mean either the 2nd or 3rd throws?
No, it means they are knocked down by THIS throw. Update accordingly.