Add titles to exercises with non-default capitalisation/punctuation (#901)
In generated READMEs, the rule for converting an exercise slug to a
human-readable form is to capitalise only the first letter of each
hyphen-separated word. For example, `lorem-ipsum` will become "Lorem
Ipsum". This is true of both the Trackler and Configlet README
generators.

Trackler's README generator (see `name` function):
https://github.com/exercism/trackler/blob/master/lib/trackler/specification.rb

Configlet's README generator (see `Name` function):
https://github.com/exercism/configlet/blob/master/track/problem_specification.go

For some exercises this is not the best possible title. For example see
http://exercism.io/exercises/haskell/rna-transcription/readme and
observe that the title is "Rna Transcription", whereas "RNA
Transcription" seems more faithful to the rules (RNA, as an initialism,
should be in all caps).

If we would like the README generator to generate these non-default
titles, it seems to make the most sense to add it in
problem-specifications. The alternative of adding the exceptions to
configlet has the disadvantage of requiring an update to configlet for
any new exercise that desires a non-default title.

The proposal is a `title` field in metadata.yml that overrides the
default capitalisation or punctuation. For those exercises for which the
default title is appropriate, it is recommend that this field be
omitted, to avoid unnecessary configuration.

Requires (and has received) confirmation from
https://github.com/exercism/configlet/issues/76 that this proposal is
appropriate on the configlet side before merging.

Requires documentation update before merging:
https://github.com/exercism/docs/pull/98

Configlet usage of the `title` field:
https://github.com/exercism/configlet/pull/85