markdown 1.1.0: Do not add <p> to list items
Compare the case of "unordered lists" vs "With a little bit of
everything":

In the former, each list item is surrounded by `<p>` tags.

In the latter, they are not, instead being surrounded by other tags that
they otherwise had.

This could plausibly be explained by the rule "If there are any tags on
list items, don't add any, otherwise do add `<p>`" but it seems weird
because we don't usually see `<p>` for list items.

On a survey of two arbitrarily chosen Markdown parsers:

* http://parsedown.org/demo
* https://daringfireball.net/projects/markdown/

Neither emits `<p>` for the list items.
It seems they were a mistake and should be removed, unless information
to the contrary is reported at
https://github.com/exercism/problem-specifications/pull/262#discussion_r144485967