crypto-square: Fix description copy and example (#929)
The final example in the crypto-square description is inconsistent with
the [source of the
exercise](http://users.csc.calpoly.edu/~jdalbey/103/Projects/ProgrammingPractice.html),
the canonical data, and with itself. This change does a few things to
correct that:

- While the encoding method is called a "square code," we're really
dealing with all sorts of rectangles, squares included. The description
gets this right up until the last example, and this change clarifies
that.

- In both the source of the exercise and the canonical data, when a
"chunk" is shorter than the length of a row, a blank space is added to
the end of that chunk. This change fixes that in the final example.

- The copy before the final example says that "spaces should be
distributed evenly across the last `n` rows," but it doesn't explain
how. This change specifies that the spaces should be added to the end of
each chunk.

Here's a related discussion:
https://github.com/exercism/problem-specifications/issues/356