Add canonical-data.json.
This commit adds a canonical-data.json to the diffie-hellman problem
(REF: exercism/problem-specifications#557).

There are irregularities in how this problem is being tested across the
current tracks:

csharp, fsharp, elixer and go all consistently test both for large and
small prime numbers test cases.  They also assume that the application
generates the private key, not the user picks as specified by the
problem description.

ecmascript and javascript, on the other hand, only use the very small
prime numbers as the test cases, and ask the user to implement a private
key generator.

The other major difference is that both javascript and ecmascript test
boundaries and verify that their implementations throw errors for
specific cases.  The other implementations do no do this (perhaps
because they are all typed?).

One last difference is that the go implementation uses hexadecimal
notation for the bit Ints.

I added the big primes as an optional set of tests given that some
tracks might like to implement this if handling big ints is one of the
major features of the track language.