Make food chain properties less ambiguous
Currently, the canonical data uses the same property to get a single verse and to get multiple verses. I think it would make sense to instead use a different property for multiple verses, aptly named `verses` (plural). Furthermore, the single verse cases used `"start verse"`  as the input property, but to me that suggests also having an `"end verse"` property, which it doesn't. I've therefore renamed the property for the single verse cases to just `"verse"`.