description.md: rm trailing spaces (#995)
One might assume that we can consider a CI check that there are no
trailing spaces, but we have to be careful because some embedded strings
that serve as example outputs are expected to have trailing spaces. The
currently only example is ocr-numbers. One might consider a CI check,
but simply exclude that file.