move "targetgrouping" to "comments"
It has been agreed that `targetgrouping` is not an input and nor is it an `expected` output.  Either it should be removed or moved to a comment.  Before making the final decision for removal lets see how we like it as a comment.  The comments are very basic and rewording may be desired by some.
Thanks @petertseng for helping to get comments to conform to schema.