travis: Choose schema based on USE_OLD_SCHEMA file
If it is present, use the old schema, else use the new schema.

This plan is proposed in:
https://github.com/exercism/problem-specifications/pull/998

Its merits are that it allows us to:

* Track progress by simply counting number of USE_OLD_SCHEMA file.
* Verify, as each PR is made, that each JSON file intending to use new
  schema does in fact validate.
  * Otherwise we have to manually run a schema check at periodic
    intervals.
* Keep compatibility with old schema, to support moving over one file at
  a time instead of all at once.

Note that old-schema.json is the same canonical-schema.json as is
currently on master.

The intention is that when all exercises use the new schema, **this
commit can/should be reverted** and then all exercises will in fact use
the new schema, which will replace the old completely.