Acronym canonical-data: Remove redundant test case
The test with phrase “PHP: Hypertext Preprocessor” and the other test
with phrase “GNU Image Manipulation Program” are testing for the same
thing. Both tests verify that only the first letter of the entirely
capitalised word is being returned.

Let’s remove the test case with phrase “PHP: Hypertext Preprocessor”
as PHP, being recursive, is ambiguous.

Also, let’s update the test description to better reflect what this
test is testing for.