complex-numbers: Move real and imaginary part cases to top
Currently, the `real` and `imaginary` test cases (which merely extract information from the complex number) are almost at the very end of the test cases. To me this does not really make sense, as I think these are the most simple and elementary test cases. This PR therefore moves these cases to the top.