two-fer: Remove assumption about track ordering
The exercise was originally made to have the same format as the old hello-world exercise:
provide a default value if none is passed, and use the passed argument when present.

This let us simplify hello-world so that it is a simple check to ensure that everything is
wired together correctly.

That said, there are languages where implementing this is non-trivial, so while we did make the
exercise to replace the old hello-world, it doesn't necessarily follow that it would always
directly follow a hello-world exercise.