react: Fold expect_callback_values into set_value
react 2.0.0

expect_callback_values is often not precise enough about which
`set_value` calls should result in which callbacks being called or not
called. "callbacks can be added and removed" is the worst offender,
expecting a `[32, 42]` from callback2 only after both `set_value` have
fired.

That means for all we know, maybe an implementation:

* magically manages to predict the future and calls the callback twice
  on the first `set_value`, with the correct value, and then zero times
  on the second `set_value`.
* calls the callback zero times on the first `set_value`, but twice
  on the second `set_value`.

To remedy this, we observe that the only time callbacks might be called
is if there was a `set_value`. So, let's explicitly associate our
callback expectations along with each `set_value`.