Add referance to 'espeak'
When I first read this README I skimmed past the `extension` section since it said "if you're on a Mac", and I'm using Linux. I'm glad I took the time to find an alternative, it was really fun, and I wanted to make sure others will be made aware of it too.

I'd be happy to receive feedback on the construction of the sentence. I'm sure it can be improved.