isbn-verifier: Crafted input to catch more incorrect algorithms (#1255)
* isbn-verifier: Crafted input to catch more incorrect algorithms

This commit adds an input that will be incorrectly validated by
algorithms that don't explicitly check the length of their inputs, as
discussed in #1199 and #993.

For example, if an algorithm only checks the first (or last) ten
digits of the input, ignoring leftovers, it will pass the current test
suite even though it doesn't correctly implement the spec.

closes #1199

* Remove old length test case

The "too long no dashes" test case will serve the same purpose, and will
allow certain solutions that don't check length (e.g. that test the first
or last ten digits of the input, or that generalise the checksum to more
digits) to look like they're working until they reach the new "too long
but contains valid isb" exercise.