book-store: Add test case to thwart list-order greedy solutions (#1297)
The current "two groups of four" test allows solutions to pass if they add books
to groups in the order in which they appear in the list, which doesn't work in
general. This test will cause those solutions to fail, because adding books 2
and 3 to the first group of four will prevent making a second one.