Major version change on the zipper exercise.
We added a an insertion method to the zipper class. The main goal is
to be able to run to run test without using an intermediate tree structure.

Now the test consists in creating a tree starting from an empty zipper
using the the atomic operations left, right, up and insert/modify/read
node values.

In the old version there were no immutability check. Immutability will
now be an important part of students goals.