sum-of-multiples-1.3.0: there are no positive multiples of 0
As I was solving this exercise on the Haskell track, Frerich pointed out
that my solution didn't terminate for `sumOfMultiples [0] 1`. This was
due to `[0,factor..limit-1]` producing an infinite stream of zeroes.
Since the only multiple of 0 is 0 itself, the result should be 0.

This change addresses 0 as a possible input factor (test case) but does
not change the wording of the exercise.