scale-generator: update test properties: one property per signature
Many tracks use automatic or semiautomatic tools to assist in
implementing an exercise for that track. These tools are very
useful, but depend on certain properties of the canonical data.

One such property is that all tests with the same "property"
field should have the same signature: the same keys in the
"input" object and the same type of "expected" object.

The scale-generator exercise used to violate that property:
all tests used the "pitches" property. This PR brings it into
compliance by separating the "chromatic" tests from the
"interval" tests. These new names, in addition to aiding
test generators, are more descriptive than "pitches".