sum-of-multiples: describe the importance of each test
sum-of-multiples 1.4.1

As stated in our guidelines, we want descriptions that tell us why each
case is being tested:
https://github.com/exercism/problem-specifications#test-data-format-canonical-datajson

The canonical data was added in
https://github.com/exercism/problem-specifications/pull/206 during a
time when descriptions were not required.

Then, placeholder descriptions were added in
aaa47eef1af516f96cdffc0aa23b63e06367c6fd and it was acknowledged at the
time that these descriptions were purely for *schema* compliance, and
not compliance with the guidelines nor a motivation for each case.

By examining the issues that led to each case being added, it is
possible to understand the descriptions.

With more cases being added to sum-of-multiples with clear descriptions,
it's time to bring the other descriptions up to speed as well.

Original exercise:
https://github.com/exercism/exercism/commit/607be68494d30487c001303b6dc56c2b7b975409

Additional cases:
https://github.com/exercism/exercism/issues/2486
https://github.com/exercism/exercism/issues/2341
https://github.com/exercism/haskell/pull/50
https://github.com/exercism/go/pull/53
https://github.com/exercism/problem-specifications/pull/896