sum-of-multiples: handle cardinality > 3 correctly (#1368)
It is possible to approach this exercises using an entirely arithmetic
approach instead of a container type. In this approach the total of
the arithmetic series of each factor must be compensated for overlaps
using the inclusion-exclusion principle; the current body of tests
handle only up to a cardinality of 3, but it is quite easy to get
something passing a cardinality of 3 without realizing it will fail
with cardinality of 4 or more. A test with 5 prime factors will
demonstrate the ability of the solution to handle the problem
generally.

Limit is 10000 to ensure 4-wise multiples are excluded and 5-wise are
included.

Closes https://github.com/exercism/problem-specifications/issues/1367