wordy: test that problems with zero operations give the correct answer
wordy 1.4.0

It's possible to have a healthy debate about whether a problem with zero
operations really should count as in scope.

One might expect that a generically-written solution would indeed accept
such a problem.

Consider an implementation that strips away all tokens except for
numbers and the four recognised operators (ignoring the intro like "What
is"). Such an implementation would now be required to distinguish
between the valid "What is 5?" versus the invalid "What is 4 cubed"
instead of simply stripping away the "cubed" and declaring the problem
invalid because it only contains the single token `4` (for that would
also declare `What is 5?` invalid)