wordy: test that various syntax errors are in fact syntax errors
By explicitly testing these we are making sure that the implementation
fails in an appropriate way, rather than failing in an inappropriate way
or erroneously giving an answer.

For example, in languages using an optional for their answer, ensuring
that [None is returned instead of panicking/aborting][1].

[1]: https://github.com/exercism/rust/pull/704#discussion_r230593072

This is beneficial because the problem is partially about building a
parser (even if very rudimentary), and it's common for parsers to have
to deal with errors like this. We may otherwise see some unfortunately
fragile parsers.