wordy: Test question `What is?` because it's too short
wordy 1.5.0

In keeping with [1.3.0][], we want to test that this fails in an
expected way rather than an unexpected way.

[1.3.0]: https://github.com/exercism/problem-specifications/pull/1383

For example, imagine a language that uses an optional as the answer. We
would want to make sure an implementation written in that language would
return `None`, rather than (say) accessing an out-of-bounds index due to
always assuming that the inputs will have at least three space-separated
elements (of the form "What is X...?").