Leap: Remove recommendation to DIY in description
Leap-year functions don't generally occur in standard libraries.

Is there a single example of a language we have that does this?

If there is one, then that would be the right way to solve this exercise
in that language. Presumably this exercise is primarily placed at the
top of core tracks. We should not encourage to re-implement standard
library functions at that point. But this is a hypothetical argument.