palindrome-products: provide more meaningful error messages
The two tests where the range is invalid (min>max) produce a bad error
message, stating "invalid input: min is X and max is Y", which is
something of a "go figure". Instead, they should specify the expectation
that has been violated, namely that min must be smaller or equal to max.