darts: add tests to disallow absolute addition (#1488)
* darts: add tests to disallow absolute addition

The current tests allow an erroneous-but-passing solution where the sum of the absolute values of the coordinates of the dart are compared to the radius thresholds (ie return 5 if `abs(x) + abs(y)` is greater than 1 but less than or equal to 5), rather than the calculated distance of the dart from the origin. Thanks to grwkremilek for the find.