markdown: unordered lists with surrounding content (#1490)
Adds a tests for unordered lists with surrounding content; many
solutions I've seen are not properly closing an unordered list when
there's content below the list's final item.

Note that there's only a single newline between the final item and the 
content right after it. 
Some Markdown parsers seem to parse this as a continuation of the line
item rather than ending the list.
The Markdown spec does not seem to clearly define this, so we
acknowledge this ambiguity.

We feel that the choice we make doesn't affect the educational value of
the exercise, but does avoid some complexity in a parser.