Reverse String: move unicode extensions to exercises
Move the unicode bonus to the exercises instead of the description. This
will allow the exercises to automatically be brought into the various
tracks.

I was originally going to use 👍🏽 as the grapheme example. But as it is
split in my terminal (Gnome Terminal) and my editor (Visual Studio
Code), I presume it would be split for most students. Displaying it
split might make it confusing.