Fix: Add missing test case that distinguishes from continue and break for loops.
1. Why is this change neccesary?
Because the current test suite allows you to use a `break` a statement inside a
for loop and having all the test pass, this can lead to some confusion between
`break` and `continue` statements that are inside a for loop.

With the current test suite one could do something like this:

```
for i, letter := range word {
		letter = unicode.ToLower(letter)
		if !unicode.IsLetter(letter) {
			break
		}
		if strings.ContainsRune(word[i+1:], letter) {
			return false
		}
	}
```

And have all the test pass.

2. How does it address the issue?
A test case was added where it won't pass if you use a `break` statement, you
need to use a `continue` statement for it to work.

3. What side effects does this change have?
It clears out the confusion between `break` and `continue`.