[Darts] Slim down generated method names (#1530)
The purpose of the "description" field is to derive identifiers for naming generated test methods, but some descriptions have tended to verbosity ending up with long identifiers (up to 150 characters).  Such long method names are harder for students to work with and also are awkward for some IDE GUIs.

Issue #1473 proposed adding a separate "identifier" field, but the consensus there was to first try improving "descriptions" to better suit "identifier" generation.    This PR attempts to slim down the generated identifiers without losing substantive information about the test.

For the longest descriptions here, I couldn't recognize the significance of...
```
dart whose coordinates sum to > 1 but whose radius to origin is <= 1
```
While considering a shorter alternative, it helped consistency to also rearranged the shorter descriptions to be more explicit about each test condition, and add a couple of tests each side of boundary conditions.

Co-Authored-By: Erik Schierboom <erik_schierboom@hotmail.com>
Co-Authored-By: Derk-Jan Karrenbeld <derk-jan+github@karrenbeld.info>