say: Tests vulnerable to octal interpretation of chunks
The tests of `say` do not catch an erroneous interpretation of chunks as
octal, which happens easily with internal zeroes. For example, my code
produced the following:

````bash
bash say.sh 1023
one thousand nineteen
````

This is due to `023 oct == 19 dec`. Unfortunately, none of the test
cases test for this (and I'm sure quite a lot of implementations in the
database are wrong because of this). These two cases catch

a) wrong interpretation with valid octal numbers
b) interpretation of chunks containing digits invalid in octal, i.e. 8
and 9.