word-search: ensure every letter of the word is checked
I thought of adding this case, but I think it's highly unlikely that any
submitted solution will make this mistake.

The most likely form of this mistake is an off-by-one mistake, where
only the first N-1 letters of the word are checked. But we already have
coverage for this in the existing tests, even if it isn't explicit.
The word "lua", if suject to this mistake, has at least one location
where it might become "lur" instead.

Since this test case has not in fact added any new coverage, I have
decided to withdraw it.