Add micro-blog exercise (#1509)
* Add micro-blog exercise

This is an exercise requiring students to truncate unicode strings.
Solves #1507

* Micro-blog: Don't assume native English speaker

Thank you @SaschaMann for the feedback and suggestion.
https://github.com/exercism/problem-specifications/pull/1509#discussion_r277131585

> I don't like that this assumes the perspective of a native English
> speaker. English is a foreign language to most of the world. Perhaps
> something along the lines of "text in most of the world's languages and
> scripts" would be a better description.

* Micro-blog: Add tests for different languages

Feedback from @SaschaMann
https://github.com/exercism/problem-specifications/pull/1509#issuecomment-485091337

> I think it would be nice to add some test cases that aren't emoji or
> English - perhaps cases with germanic umlauts, cyrillic and/or greek
> letters, historic scripts etc. - because that's one of the main uses
> and goals of unicode.

I've added German, Bulgarian, and Greek examples. All of them have
non-English characters.

None of these characters use multiple UTF-16 codepoints. As such, if you
use a UTF-8 programming language you may first have trouble with the
German example, but if you use a UTF-16 language you will probably first
have trouble at the Emoji example.

I chose not to add an example with historic scripts, because I'm not
aware of any that display nicely in my terminal or text-editor. Perhaps
in future some could be added.

I wanted another example that would be problematic in UTF-16, so I added
a poker hand example using playing cards.

* Micro-blog: Add German truncated example

Comically, it goes from "bear carpet" to "beards".

@SaschaMann, thank you for finding the example for me:
https://github.com/exercism/problem-specifications/pull/1509#issuecomment-485273006

* Micro-blog: Add longer maths example

Empty set is a proper subset of the natural numbers which is a proper
subset of the integers, which is a proper subset of the rational numbers
which is a proper subset of the reals which is a proper subset of the
complex numbers.

It remains true when truncated which is quite nice