topics: suggestion to add errors
This is a bit of a Go thing.

First of all there are no `exceptions` in Go. So we cannot use the tag `exception_handling` as that would be wrong. In Go we have `errors` (and `panics`).
Secondly, errors are values in Go. So they'd be a data type.

We could also add `error_handling` in `Basic Concepts` instead.